pyproject.toml
[project]
name = "deepseek-ocr-eval"
version = "0.1.0"
description = "Evaluation tools for DeepSeek-OCR compression ratio and accuracy"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=10.0.0",
    "python-levenshtein>=0.21.0",
    "torch>=2.0.0",
    "transformers>=4.40.0,<4.47.0",
    "accelerate>=0.20.0",
    "einops>=0.6.0",
    "timm>=0.9.0",
    "addict>=2.4.0",
    "easydict>=1.13",
    "datasets>=4.4.1",
]

[project.scripts]
deepseek-ocr-eval = "eval:main"

[dependency-groups]
dev = [
    "ruff>=0.14.8",
]
